Move plot function to plot.py
